sm6225-common: device: Import android.hardware.media.c2 to FCM
* Also build missing c2 libs

hwservicemanager: Since
android.hardware.media.c2@1.0::IComponentStore/default is not
registered, trying to start it as a lazy HAL.
CodecServiceRegistrant: Preferred Codec2 store is defaulted to
"software".
CodecServiceRegistrant: Software Codec2 service created and registered.
libc    : Unable to set property "ctl.interface_start" to
"android.hardware.media.c2@1.0::IComponentStore/default": error code:
0x20
hwservicemanager: Tried to start
android.hardware.media.c2@1.0::IComponentStore/default as a lazy
service, but was unable to. Usually this happens when a service is not
installed, but if the service is intended to be used as a lazy service,
then it may be configured incorrectly.

Change-Id: I348d0929ebab99f27e75abfee63168499fe7ad1a